requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy<2.0.0
pandas>=1.1.0
tqdm>=4.50.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.23.0
open3d>=0.12.0
pyyaml>=5.3.0 